global: refactor and normalize controllers
Controllers have been moved to src/main/java/controller and belongs to
the controller namespace. fxml files have been moved to
src/main/resources/fxml.

Updated gradle configuration to build using Side_Panel.fxml as the entry
point.
