+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'com.github.sherter.google-java-format' version '0.8'
 }
 
-mainClassName = 'Main'
+mainClassName = 'controller.MainController'
 
 dependencies {
 }
