+++ src/main/java/controller/CourseMainPage.java
@@ -1,4 +1,4 @@
-package views;
+package controller;
 
 import javafx.fxml.FXML;
