+++ src/main/java/controller/CourseOverviewPage.java
@@ -1,4 +1,4 @@
-package GUI;
+package controller;
 
 public class CourseOverviewPage {
 }
