+++ src/main/java/controller/MainController.java
@@ -1,0 +1,25 @@
+package controller;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class MainController extends Application {
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage stage) throws IOException {
+        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("fxml/Side_Panel.fxml"));
+        Scene scene = new Scene(root, 300, 275);
+        stage.setTitle("Min5a");
+        stage.setScene(scene);
+        stage.show();
+    }
+}
