+++ src/main/java/controller/StaticsPage.java
@@ -1,3 +1,5 @@
+package controller;
+
 import javafx.fxml.FXML;
 import javafx.scene.text.Text;
