+++ src/main/resources/fxml/CourseMainPage.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="400.0" prefWidth="510.0" stylesheets="@min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="views.CourseMainPage">
+<AnchorPane prefHeight="400.0" prefWidth="510.0" stylesheets="@../min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.CourseMainPage">
    <children>
       <TextArea fx:id="CourseName" editable="false" layoutX="14.0" layoutY="14.0" prefHeight="50.0" prefWidth="322.0" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="170.0" AnchorPane.topAnchor="15.0" />
       <SplitPane dividerPositions="0.5" layoutX="15.0" layoutY="74.0" prefHeight="235.0" prefWidth="325.0" AnchorPane.bottomAnchor="46.0" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="170.0">
