+++ src/main/resources/fxml/CourseOverviewPage.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.text.Text?>
 
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="510.0" stylesheets="@min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GUI.CourseOverviewPage">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="510.0" stylesheets="@../min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.CourseOverviewPage">
    <children>
       <Text layoutX="17.0" layoutY="56.0" strokeType="OUTSIDE" strokeWidth="0.0" styleClass="header" text="Pågående Kurser" AnchorPane.leftAnchor="19.0" />
       <ScrollPane layoutX="18.0" layoutY="66.0" prefHeight="107.0" prefWidth="450.0" style="-fx-background-color: rgba(255,255,255,0)" AnchorPane.leftAnchor="18.0" AnchorPane.rightAnchor="41.60000000000002">
