+++ src/main/resources/fxml/CoursePanelItem.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="105.0" prefWidth="142.0" style="-fx-background-radius: 6; -fx-border-width: .5; -fx-border-radius: 6; -fx-border-color: #999999;" styleClass="panel-background3" stylesheets="@min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GUI.CoursePanelItem">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="105.0" prefWidth="142.0" style="-fx-background-radius: 6; -fx-border-width: .5; -fx-border-radius: 6; -fx-border-color: #999999;" styleClass="panel-background3" stylesheets="@../min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.CoursePanelItem">
    <children>
       <ImageView fx:id="courseImage" fitHeight="75.0" fitWidth="128.0" layoutX="7.0" layoutY="7.0" pickOnBounds="true" preserveRatio="true">
          <image>
