+++ src/main/resources/fxml/StaticsPage.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" stylesheets="@views/min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="StaticsPage">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" stylesheets="@views/min5a.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.StaticsPage">
    <children>
       <AnchorPane layoutX="136.0" layoutY="205.0" prefHeight="160.0" prefWidth="250.0" styleClass="panel-background3" stylesheets="@views/min5a.css">
          <children>
