+++ src/main/resources/logo.png
