+++ src/main/resources/min5a.properties
